+++ Civi/ActionProvider/Action/Communication/SendPdfByEmail.php
@@ -72,6 +72,9 @@ class SendPdfByEmail extends AbstractAction {
       $mailer->setSenderContactId($parameters->getParameter('sender_contact_id'), true, false);
     }
 
+    if ($parameters->getParameter('location_type_id')) {
+      $mailer->setLocationTypeId($parameters->getParameter('location_type_id'));
+    }
     if ($parameters->doesParameterExists('case_id')) {
       $mailer->setCaseId($parameters->getParameter('case_id'));
     }
@@ -178,6 +181,7 @@ class SendPdfByEmail extends AbstractAction {
     $attachments->setDescription(E::ts('Give either the path to the file or the File ID in the CiviCRM database.'));
     return new SpecificationBag(array(
       new Specification('contact_id', 'Integer', E::ts('Receiver Contact ID'), true),
+      new Specification('location_type_id', 'Integer', E::ts('Location Type'), false),
       new Specification('alternative_recipient', 'String', E::ts('Alternative Recipient Email'), false),
       new Specification('subject', 'String', E::ts('Subject'), true),
       new Specification('body_html', 'String', E::ts('HTML Body'), true),