+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.106 (not yet released)
 ------------
 
+* Added location type to Send E-mail and Send PDF by Email actions. With a fall back to primary email address.
+
 Version 1.105
 ------------